Initial implementation of module imports for simple EXTENDS
Currently, the idea is that imported modules will be searched for based on the URL of the given specification path argument. So, if a spec is loaded from `<url>/spec.tla`, then an imported module M will be searched for and fetched from `<url>/M.tla`. This is an attempt to emulate the module import behavior of TLC i.e., modules that exist in the same directory as a root module should be automatically discovered. Eventually this could be generalized to fetch modules from additional, user-specified locations.