Add new logic and tests to account for name clashes between operator arguments defined before other declarations (e.g. CONSTANT or VARIABLE declarations)
Basically, if an operator is defined before a VARIABLE declaration, then its "current context" does not include that variable name, and so an operator argument name won't actually clash with that VARIABLE name. So, it's fine for an operator argument and a VARIABLE to share the same name in this case. But, we just need to make sure that the interpreter properly recognizes this situation and treats the operator argument appropriately i.e. in the "current context" of where the operator was originally defined.