Refactor handling of definitions and contexts throughout the interpreter
Restructure of how we store and look up definitions. Namely, every expression evaluation or definition is now associated with a "current context", which essentially defines the set of definitions that were in scope at the time that unique definition was first defined. Along with this, we also move to a model where all definitions that were ever defined in any module are assigned globally unique identifiers, which can be used to reference them in a given scope/context of evaluation.

Also updates and adds various tests for exercising this new behavior more thoroughly.